Create PROFIGY_CS_02.PY
This code uses NumPy for scrambling and unscrambling the pixels of the image. When you click "Encrypt Image," the pixels of the selected image will be scrambled and saved as a new encrypted image. When you click "Decrypt Image," the pixels of the encrypted image will be unscrambled, and the original image will be restored